Fix PCE Mouse on multitap
Direct connection hs a very short CLR pulse, and the CLR signal remains low for the majority of the time. It was for this reason that the clock.pio program waited for the positive edge (0->1) of the CLR signal.

However, on the multitap, the CLR signal is on all of the time except when that particular port is enabled. For this reason, the clock.pio needs to be changed to wait for the negative edge (1->0)  of the CLR signal.